Refactor common save methods to BaseInteractor
