+++ opensrp-reveal/src/main/java/org/smartregister/reveal/contract/ListTaskContract.java
@@ -59,5 +59,9 @@ public interface ListTaskContract {
         void onStructureAdded(Feature feature, JSONArray featureCoordinates);
 
         void onFormSaveFailure(String eventType);
+
+        void onCardDetailsFetched(CardDetails cardDetails);
+
+        void onInterventionFormDetailsFetched(CardDetails finalCardDetails);
     }
 }
