+++ opensrp-reveal/src/main/java/org/smartregister/reveal/interactor/BaseInteractor.java
@@ -48,6 +48,11 @@ import static com.cocoahero.android.geojson.Geometry.JSON_COORDINATES;
 import static org.smartregister.reveal.util.Constants.BEDNET_DISTRIBUTION_EVENT;
 import static org.smartregister.reveal.util.Constants.DETAILS;
 import static org.smartregister.reveal.util.Constants.Intervention.BEDNET_DISTRIBUTION;
+import static org.smartregister.reveal.util.Constants.Intervention.IRS;
+import static org.smartregister.reveal.util.Constants.Intervention.LARVAL_DIPPING;
+import static org.smartregister.reveal.util.Constants.Intervention.MOSQUITO_COLLECTION;
+import static org.smartregister.reveal.util.Constants.JsonForm.ENCOUNTER_TYPE;
+import static org.smartregister.reveal.util.Constants.LARVAL_DIPPING_EVENT;
 import static org.smartregister.reveal.util.Constants.METADATA;
 import static org.smartregister.reveal.util.Constants.MOSQUITO_COLLECTION_EVENT;
 import static org.smartregister.reveal.util.Constants.REGISTER_STRUCTURE_EVENT;
@@ -102,14 +107,11 @@ public abstract class BaseInteractor {
         try {
             JSONObject jsonForm = new JSONObject(json);
             String encounterType = jsonForm.getString(JsonForm.ENCOUNTER_TYPE);
-            if (SPRAY_EVENT.equals(encounterType)) {
-                saveSprayForm(jsonForm);
+            if (SPRAY_EVENT.equals(encounterType) || MOSQUITO_COLLECTION_EVENT.equals(encounterType)
+                    || LARVAL_DIPPING_EVENT.equals(encounterType) || BEDNET_DISTRIBUTION_EVENT.equals(encounterType)) {
+                saveMosquitoInterventionForm(jsonForm);
             } else if (REGISTER_STRUCTURE_EVENT.equals(encounterType)) {
                 saveRegisterStructureForm(jsonForm);
-            } else if (MOSQUITO_COLLECTION_EVENT.equals(encounterType)) {
-                saveMosquitoCollectionForm(jsonForm);
-            } else if (BEDNET_DISTRIBUTION_EVENT.equals(encounterType)) {
-                saveBednetDistributionForm(jsonForm);
             }
         } catch (Exception e) {
             Log.e(TAG, "Error saving Json Form data", e);
@@ -120,6 +122,7 @@ public abstract class BaseInteractor {
         String entityId = getString(jsonForm, ENTITY_ID);
         JSONArray fields = JsonFormUtils.fields(jsonForm);
         JSONObject metadata = getJSONObject(jsonForm, METADATA);
+
         FormTag formTag = new FormTag();
         formTag.providerId = sharedPreferences.fetchRegisteredANM();
         formTag.locationId = sharedPreferences.fetchDefaultLocalityId(formTag.providerId);
@@ -130,29 +133,44 @@ public abstract class BaseInteractor {
         eventJson.put(DETAILS, getJSONObject(jsonForm, DETAILS));
         eventClientRepository.addEvent(entityId, eventJson);
         return gson.fromJson(eventJson.toString(), org.smartregister.domain.db.Event.class);
-
-
     }
 
-    private void saveSprayForm(JSONObject jsonForm) {
+    private void saveMosquitoInterventionForm(JSONObject jsonForm) {
+        String interventionType = null;
+        String encounterType = null;
+        try {
+            encounterType = jsonForm.getString(ENCOUNTER_TYPE);
+            if (encounterType.equals(SPRAY_EVENT)) {
+                interventionType = IRS;
+            } else if (encounterType.equals(MOSQUITO_COLLECTION_EVENT)) {
+                interventionType = MOSQUITO_COLLECTION;
+            } else if (encounterType.equals(LARVAL_DIPPING_EVENT)) {
+                interventionType = LARVAL_DIPPING;
+            } else if (encounterType.equals(BEDNET_DISTRIBUTION_EVENT)) {
+                interventionType = BEDNET_DISTRIBUTION;
+            }
+        } catch (JSONException e) {
+            Log.e(TAG, e.getStackTrace().toString());
+        }
+
+        final String finalInterventionType = interventionType;
+        final String finalEncounterType = encounterType;
         Runnable runnable = new Runnable() {
             @Override
             public void run() {
                 try {
-                    org.smartregister.domain.db.Event event = saveEvent(jsonForm, SPRAY_EVENT, STRUCTURE);
+                    org.smartregister.domain.db.Event event = saveEvent(jsonForm, finalEncounterType, STRUCTURE);
                     clientProcessor.processClient(Collections.singletonList(new EventClient(event, null)), true);
                     appExecutors.mainThread().execute(new Runnable() {
                         @Override
                         public void run() {
                             String businessStatus = clientProcessor.calculateBusinessStatus(event);
-                            presenterCallBack.onFormSaved(event.getBaseEntityId(),
-                                    Task.TaskStatus.COMPLETED, businessStatus, Intervention.IRS);
-
+                            presenterCallBack.onFormSaved(event.getBaseEntityId(), Task.TaskStatus.COMPLETED, businessStatus, finalInterventionType);
                         }
                     });
                 } catch (JSONException e) {
                     Log.e(TAG, "Error saving spraye", e);
-                    presenterCallBack.onFormSaveFailure(SPRAY_EVENT);
+                    presenterCallBack.onFormSaveFailure(finalEncounterType);
                 }
             }
         };
@@ -243,51 +261,6 @@ public abstract class BaseInteractor {
             }
         };
 
-        appExecutors.diskIO().execute(runnable);
-    }
-
-    private void saveMosquitoCollectionForm(JSONObject jsonForm) {
-        Runnable runnable = new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    org.smartregister.domain.db.Event event = saveEvent(jsonForm, MOSQUITO_COLLECTION_EVENT, STRUCTURE);
-                    clientProcessor.processClient(Collections.singletonList(new EventClient(event, null)), true);
-                    appExecutors.mainThread().execute(new Runnable() {
-                        @Override
-                        public void run() {
-                            String businessStatus = clientProcessor.calculateBusinessStatus(event);
-                            presenterCallBack.onFormSaved(event.getBaseEntityId(), Task.TaskStatus.COMPLETED, businessStatus, Intervention.MOSQUITO_COLLECTION);
-                        }
-                    });
-                } catch (Exception e) {
-                    Log.e(TAG, "Error saving mosquito collection point data");
-                }
-            }
-        };
-        appExecutors.diskIO().execute(runnable);
-    }
-
-
-    private void saveBednetDistributionForm(JSONObject jsonForm) {
-        Runnable runnable = new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    org.smartregister.domain.db.Event event = saveEvent(jsonForm, BEDNET_DISTRIBUTION_EVENT, STRUCTURE);
-                    clientProcessor.processClient(Collections.singletonList(new EventClient(event, null)), true);
-                    appExecutors.mainThread().execute(new Runnable() {
-                        @Override
-                        public void run() {
-                            String businessStatus = clientProcessor.calculateBusinessStatus(event);
-                            presenterCallBack.onFormSaved(event.getBaseEntityId(), Task.TaskStatus.COMPLETED, businessStatus, BEDNET_DISTRIBUTION);
-                        }
-                    });
-                } catch (Exception e) {
-                    Log.e(TAG, "Error saving bednet distribution point data");
-                }
-            }
-        };
         appExecutors.diskIO().execute(runnable);
     }
 }
