+++ opensrp-reveal/src/main/java/org/smartregister/reveal/interactor/ListTaskInteractor.java
@@ -1,63 +1,34 @@
 package org.smartregister.reveal.interactor;
 
-import android.content.Context;
 import android.util.Log;
 
-import com.google.gson.JsonArray;
 import com.mapbox.geojson.Feature;
 
 import net.sqlcipher.Cursor;
 import net.sqlcipher.database.SQLiteDatabase;
 
-import org.joda.time.DateTime;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.smartregister.clientandeventmodel.Event;
 import org.smartregister.domain.Location;
-import org.smartregister.domain.LocationProperty;
 import org.smartregister.domain.Task;
-import org.smartregister.domain.db.EventClient;
-import org.smartregister.domain.tag.FormTag;
-import org.smartregister.repository.BaseRepository;
-import org.smartregister.reveal.R;
 import org.smartregister.reveal.application.RevealApplication;
 import org.smartregister.reveal.contract.ListTaskContract;
 import org.smartregister.reveal.model.CardDetails;
 import org.smartregister.reveal.model.MosquitoHarvestCardDetails;
 import org.smartregister.reveal.model.SprayCardDetails;
-import org.smartregister.reveal.util.Constants;
 import org.smartregister.reveal.util.Constants.GeoJSON;
 import org.smartregister.reveal.util.GeoJsonUtils;
-import org.smartregister.reveal.util.PreferencesUtil;
 import org.smartregister.reveal.util.Utils;
-import org.smartregister.util.JsonFormUtils;
 
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.UUID;
 
-import static com.cocoahero.android.geojson.Geometry.JSON_COORDINATES;
-import static org.smartregister.reveal.util.Constants.BusinessStatus.NOT_VISITED;
-import static org.smartregister.reveal.util.Constants.DETAILS;
 import static org.smartregister.reveal.util.Constants.Intervention.IRS;
-import static org.smartregister.reveal.util.Constants.Intervention.IRS_VISIT;
 import static org.smartregister.reveal.util.Constants.Intervention.LARVAL_DIPPING;
 import static org.smartregister.reveal.util.Constants.Intervention.MOSQUITO_COLLECTION;
-import static org.smartregister.reveal.util.Constants.JsonForm.ENCOUNTER_TYPE;
-import static org.smartregister.reveal.util.Constants.LARVAL_DIPPING_EVENT;
-import static org.smartregister.reveal.util.Constants.METADATA;
-import static org.smartregister.reveal.util.Constants.MOSQUITO_COLLECTION_EVENT;
-import static org.smartregister.reveal.util.Constants.REGISTER_STRUCTURE_EVENT;
-import static org.smartregister.reveal.util.Constants.SPRAY_EVENT;
-import static org.smartregister.reveal.util.Constants.STRUCTURE;
 import static org.smartregister.reveal.util.Constants.Tables.LARVAL_DIPPINGS_TABLE;
 import static org.smartregister.reveal.util.Constants.Tables.MOSQUITO_COLLECTIONS_TABLE;
-import static org.smartregister.util.JsonFormUtils.ENTITY_ID;
-import static org.smartregister.util.JsonFormUtils.getJSONObject;
-import static org.smartregister.util.JsonFormUtils.getString;
 
 /**
  * Created by samuelgithengi on 11/27/18.
@@ -199,164 +170,6 @@ public class ListTaskInteractor extends BaseInteractor {
             return null;
         }
         return featureCollection;
-    }
-
-    public void saveJsonForm(String json) {
-        try {
-            JSONObject jsonForm = new JSONObject(json);
-            String encounterType = jsonForm.getString(ENCOUNTER_TYPE);
-            if (SPRAY_EVENT.equals(encounterType) || MOSQUITO_COLLECTION_EVENT.equals(encounterType) || LARVAL_DIPPING_EVENT.equals(encounterType)) {
-                saveMosquitoInterventionForm(jsonForm);
-            } else if (REGISTER_STRUCTURE_EVENT.equals(encounterType)) {
-                saveRegisterStructureForm(jsonForm);
-            }
-        } catch (Exception e) {
-            Log.e(TAG, "Error saving Json Form data", e);
-        }
-    }
-
-
-    private org.smartregister.domain.db.Event saveEvent(JSONObject jsonForm, String encounterType, String bindType) throws JSONException {
-        String entityId = getString(jsonForm, ENTITY_ID);
-        JSONArray fields = JsonFormUtils.fields(jsonForm);
-        JSONObject metadata = getJSONObject(jsonForm, METADATA);
-
-        FormTag formTag = new FormTag();
-        formTag.providerId = sharedPreferences.fetchRegisteredANM();
-        formTag.locationId = sharedPreferences.fetchDefaultLocalityId(formTag.providerId);
-        formTag.teamId = sharedPreferences.fetchDefaultTeamId(formTag.providerId);
-        formTag.team = sharedPreferences.fetchDefaultTeam(formTag.providerId);
-        Event event = JsonFormUtils.createEvent(fields, metadata, formTag, entityId, encounterType, bindType);
-        JSONObject eventJson = new JSONObject(gson.toJson(event));
-        eventJson.put(DETAILS, getJSONObject(jsonForm, DETAILS));
-        eventClientRepository.addEvent(entityId, eventJson);
-        return gson.fromJson(eventJson.toString(), org.smartregister.domain.db.Event.class);
-    }
-
-    private void saveRegisterStructureForm(JSONObject jsonForm) {
-        Runnable runnable = new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    jsonForm.put(ENTITY_ID, UUID.randomUUID().toString());
-                    org.smartregister.domain.db.Event event = saveEvent(jsonForm, REGISTER_STRUCTURE_EVENT, STRUCTURE);
-                    com.cocoahero.android.geojson.Feature feature = new com.cocoahero.android.geojson.Feature(new JSONObject(event.findObs(null, false, "structure").getValue().toString()));
-                    DateTime now = new DateTime();
-                    Location structure = new Location();
-                    structure.setId(event.getBaseEntityId());
-                    structure.setType(feature.getType());
-                    org.smartregister.domain.Geometry geometry = new org.smartregister.domain.Geometry();
-                    geometry.setType(org.smartregister.domain.Geometry.GeometryType.valueOf(feature.getGeometry().getType().toUpperCase()));
-                    JsonArray coordinates = new JsonArray();
-                    JSONArray featureCoordinates = feature.getGeometry().toJSON().getJSONArray(JSON_COORDINATES);
-                    coordinates.add(Double.parseDouble(featureCoordinates.get(0).toString()));
-                    coordinates.add(Double.parseDouble(featureCoordinates.get(1).toString()));
-                    geometry.setCoordinates(coordinates);
-                    structure.setGeometry(geometry);
-                    LocationProperty properties = new LocationProperty();
-                    String structureType = event.findObs(null, false, Constants.JsonForm.STRUCTURE_TYPE).getValue().toString();
-                    properties.setType(structureType);
-                    properties.setEffectiveStartDate(now);
-                    properties.setParentId(operationalAreaId);
-                    properties.setStatus(LocationProperty.PropertyStatus.PENDING_REVIEW);
-                    properties.setUid(UUID.randomUUID().toString());
-                    structure.setProperties(properties);
-                    structure.setSyncStatus(BaseRepository.TYPE_Created);
-                    structureRepository.addOrUpdate(structure);
-
-                    Task task = new Task();
-                    task.setIdentifier(UUID.randomUUID().toString());
-                    task.setCampaignIdentifier(PreferencesUtil.getInstance().getCurrentCampaignId());
-                    task.setGroupIdentifier(operationalAreaId);
-                    task.setStatus(Task.TaskStatus.READY);
-                    task.setBusinessStatus(NOT_VISITED);
-                    task.setPriority(3);
-                    Context applicationContext = RevealApplication.getInstance().getApplicationContext();
-                    if (Constants.StructureType.RESIDENTIAL.equals(structureType)) {
-                        task.setCode(IRS);
-                        task.setDescription(applicationContext.getString(R.string.irs_task_description));
-                        task.setFocus(IRS_VISIT);
-                    } else if (Constants.StructureType.MOSQUITO_COLLECTION_POINT.equals(structureType)) {
-                        task.setCode(MOSQUITO_COLLECTION);
-                        task.setDescription(applicationContext.getString(R.string.mosquito_collection_task_description));
-                        task.setFocus(MOSQUITO_COLLECTION);
-                    } else if (Constants.StructureType.LARVAL_BREEDING_SITE.equals(structureType)) {
-                        task.setCode(LARVAL_DIPPING);
-                        task.setDescription(applicationContext.getString(R.string.larval_dipping_task_description));
-                        task.setFocus(LARVAL_DIPPING);
-                    }
-                    task.setForEntity(structure.getId());
-                    task.setExecutionStartDate(now);
-                    task.setAuthoredOn(now);
-                    task.setLastModified(now);
-                    task.setOwner(event.getProviderId());
-                    task.setSyncStatus(BaseRepository.TYPE_Created);
-                    taskRepository.addOrUpdate(task);
-                    appExecutors.mainThread().execute(new Runnable() {
-                        @Override
-                        public void run() {
-                            Map<String, String> taskProperties = new HashMap<>();
-                            taskProperties.put(Constants.Properties.TASK_IDENTIFIER, task.getIdentifier());
-                            taskProperties.put(Constants.Properties.TASK_BUSINESS_STATUS, task.getBusinessStatus());
-                            taskProperties.put(Constants.Properties.TASK_STATUS, task.getStatus().name());
-                            taskProperties.put(Constants.Properties.TASK_CODE, task.getCode());
-                            taskProperties.put(Constants.Properties.LOCATION_UUID, structure.getProperties().getUid());
-                            taskProperties.put(Constants.Properties.LOCATION_VERSION, structure.getProperties().getVersion() + "");
-                            taskProperties.put(Constants.Properties.LOCATION_TYPE, structure.getProperties().getType());
-                            structure.getProperties().setCustomProperties(taskProperties);
-                            presenterCallBack.onStructureAdded(Feature.fromJson(gson.toJson(structure)), featureCoordinates);
-                        }
-                    });
-                } catch (JSONException e) {
-                    Log.e(TAG, "Error saving new Structure", e);
-                    presenterCallBack.onFormSaveFailure(REGISTER_STRUCTURE_EVENT);
-                }
-            }
-        };
-
-        appExecutors.diskIO().execute(runnable);
-    }
-
-
-    private void saveMosquitoInterventionForm(JSONObject jsonForm) {
-        String interventionType = null;
-        String encounterType = null;
-        try {
-            encounterType = jsonForm.getString(ENCOUNTER_TYPE);
-            if (encounterType.equals(SPRAY_EVENT)) {
-                interventionType = IRS;
-            } else if (encounterType.equals(MOSQUITO_COLLECTION_EVENT)) {
-                interventionType = MOSQUITO_COLLECTION;
-            } else if (encounterType.equals(LARVAL_DIPPING_EVENT)) {
-                interventionType = LARVAL_DIPPING;
-            }
-        } catch (JSONException e) {
-            Log.e(TAG, e.getStackTrace().toString());
-        }
-
-        final String finalInterventionType = interventionType;
-        final String finalEncounterType = encounterType;
-        Runnable runnable = new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    org.smartregister.domain.db.Event event = saveEvent(jsonForm, finalEncounterType, STRUCTURE);
-                    clientProcessor.processClient(Collections.singletonList(new EventClient(event, null)), true);
-                    appExecutors.mainThread().execute(new Runnable() {
-                        @Override
-                        public void run() {
-                            String businessStatus = clientProcessor.calculateBusinessStatus(event);
-                            presenterCallBack.onFormSaved(event.getBaseEntityId(), Task.TaskStatus.COMPLETED, businessStatus, finalInterventionType);
-                        }
-                    });
-                } catch (JSONException e) {
-                    Log.e(TAG, "Error saving spraye", e);
-                    presenterCallBack.onFormSaveFailure(finalEncounterType);
-                }
-            }
-        };
-
-        appExecutors.diskIO().execute(runnable);
     }
 
     private ListTaskContract.Presenter getPresenter() {
